move azure adapter to own validation package
