+++ pkg/controller/provider/azure-private/factory.go
@@ -5,12 +5,12 @@
 package azureprivate
 
 import (
-	"github.com/gardener/external-dns-management/pkg/controller/provider/azure/utils"
+	"github.com/gardener/external-dns-management/pkg/controller/provider/azure/validation"
 	"github.com/gardener/external-dns-management/pkg/controller/provider/compound"
 	"github.com/gardener/external-dns-management/pkg/dns/provider"
 )
 
-const TYPE_CODE = utils.ProviderTypeAzurePrivateDNS
+const TYPE_CODE = validation.ProviderTypeAzurePrivateDNS
 
 var rateLimiterDefaults = provider.RateLimiterOptions{
 	Enabled: true,
@@ -18,7 +18,7 @@ var rateLimiterDefaults = provider.RateLimiterOptions{
 	Burst:   10,
 }
 
-var Factory = provider.NewDNSHandlerFactory(NewHandler, utils.NewAdapter(TYPE_CODE)).
+var Factory = provider.NewDNSHandlerFactory(NewHandler, validation.NewAdapter(TYPE_CODE)).
 	SetGenericFactoryOptionDefaults(provider.GenericFactoryOptionDefaults.SetRateLimiterOptions(rateLimiterDefaults))
 
 func init() {
