+++ pkg/controller/provider/azure/validation/adapter.go
@@ -2,7 +2,7 @@
 //
 // SPDX-License-Identifier: Apache-2.0
 
-package utils
+package validation
 
 import (
 	"fmt"
